External wrapper / API moved to PyroCommon
