+++ DeadlyWeapons/DeadlyWeapons.csproj
@@ -54,7 +54,6 @@
     <ItemGroup>
         <Compile Include="DFunctions\ConsoleCommands.cs" />
         <Compile Include="DFunctions\VersionChecker.cs" />
-        <Compile Include="DFunctions\Wrapper.cs" />
         <Compile Include="Main.cs" />
         <Compile Include="Modules\CustomPullover.cs" />
         <Compile Include="Modules\Panic.cs" />
