+++ PyroCommon/Main.cs
@@ -6,7 +6,7 @@ namespace PyroCommon;
 
 public class Main
 {
-    internal static readonly Func<string, bool> IsLoaded = plugName =>
+    private static readonly Func<string, bool> IsLoaded = plugName =>
         Functions.GetAllUserPlugins().Any(assembly => assembly.GetName().Name.Equals(plugName));
-    internal static bool UsingUB { get; set; } = IsLoaded("UltimateBackup");
+    internal static bool UsingUB { get; } = IsLoaded("UltimateBackup");
 }
