+++ SuperCallouts/SuperCallouts.csproj
@@ -83,7 +83,6 @@
         <Compile Include="SimpleFunctions\ConsoleCommands.cs" />
         <Compile Include="SimpleFunctions\PulloverSpots.cs" />
         <Compile Include="SimpleFunctions\VersionChecker.cs" />
-        <Compile Include="SimpleFunctions\Wrapper.cs" />
     </ItemGroup>
     <ItemGroup>
         <Content Include="SuperCallouts.ini" />
